Play Screen Looks Nice Also Map Broke
Make the 'play' screen look more like in the prototype.
However, the 'map' screen seems to break because it can't find the 'id' of it.
It seems like something about the map screen not being exported properly.
If you switch to the 'value passing' branch, then back to the master branch without the app reloading then it sometimes still works. \shrug